Refactored to separate DA pipeline from data loading utils
